Implemented fix to issue #28
Issue #28 first reported by David Rosnick via email and involved a Java reflections exception that was unhandled.  The underlying issue was in the way that the insheetUrlToVars method was being called by the rowval subroutine in jsonio.  No value for the stubname parameter was being passed through to the JVM. This has been fixed on line 87 of jsonio.ado by adding the default stubname when calling the Java plugin.